First round through the feesbacks
(11) "Browse all news" & "Our plans in details" look ugly on mobile
(2) The videos explore build experiment in mission and video next to portals appear as video on the whole screen when scrolling on the mobile
(3) White space on the right and no space on the left.
(5) The gray footer is too large on mobile
(9) Inconsistent spacing.
(21) resources button is not using sanity text